RoleService fix

+++ core/src/main/java/edu/wmi/dpri/przychodnia/server/usermanagement/service/RoleService.java
@@ -65,7 +65,7 @@ public class RoleService {
 
     @Transactional
     public void unassignRole(User user, Role role) {
-    	Role roleFromRepo = roleRepository.findOne(role.getId());
+    	Role roleFromRepo = roleRepository.findOne(role.getId().longValue());
         roleFromRepo.getUsers().removeIf(u -> u.getId().equals(user.getId()));
         roleRepository.save(roleFromRepo);
     }
